+++ share/rocm/cmake/ROCMConfigVersion.cmake
@@ -2,7 +2,6 @@
 # Copyright (C) 2023 Advanced Micro Devices, Inc.
 # ######################################################################################################################
 
-message(STATUS
-    "Use of find_package(ROCM) will be deprecated soon, please switch to find_package(ROCmCMakeBuildTools)."
-)
-find_package(ROCmCMakeBuildTools HINTS "${CMAKE_CURRENT_LIST_DIR}")
+list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_LIST_DIR}/../../rocmcmakebuildtools/cmake)
+
+include(ROCmCMakeBuildToolsConfigVersion)