+++ test/pass/rename-compatibility.cmake
@@ -2,7 +2,8 @@
 # Copyright (C) 2023 Advanced Micro Devices, Inc.
 # ######################################################################################################################
 
-message(STATUS
-    "Use of find_package(ROCM) will be deprecated soon, please switch to find_package(ROCmCMakeBuildTools)."
-)
-find_package(ROCmCMakeBuildTools HINTS "${CMAKE_CURRENT_LIST_DIR}")
+install_dir(${TEST_DIR}/libsimplecompat TARGETS package)
+test_check_package(
+    NAME simple
+    HEADER simple.h
+    TARGET simple)